tests/requirements.txt
# Testing dependencies for Workato SDK Documentation Mirror
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-timeout>=2.1.0
responses>=0.23.0
requests-mock>=1.11.0
freezegun>=1.2.0
